Created Modal Pop Up for Modifying Account
Next to each attribute of a user's account, there now exists an edit button. When a user clicks on one of these buttons to edit a value, there will be a pop up prompting the user to fill in new data. Actual communication with the back end still needs to happen though.