Added JWT Validity Check and Altered Modify Value Form
When the user loads the page, the app will verify that the jwt stored is still valid (by seeing if it has expired). If the jwt is not valid, app clears the user store.
Changed the structure of the modify modal such that when the user hits submit, it is now the modal's responsibility to send the modify request to the server. This is so that if the user entered in their password wrong or they entered in bad input, the modal will tell them.