Infrastructure Update
When page is reloaded, before anything is rendered in index.html, the app will now look at the user stored in localstorage and determine if they are logged in. This way, when user information is updated, the user can now see those changes in real time once they hit the modify button. Furthermore, the login form now reloads the page when a user logs in, it no longer needs to grab user information and store it.