Added Comments/Removed Print Statements/Fixed Bugs
Added comments to every function in the front end code (hopefully I covered them all). Also deleted all unnecessary console logs in the code.
One bug that I fixed was that templates could not be read properly when searching for them. This was fixed by adding searchTemp to the props.